backend/pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "csre"
version = "0.1.0"
description = "FastAPI backend scaffold for the Cycle-Safe Referral Engine."
requires-python = ">=3.10"
dependencies = [
  "alembic>=1.14.1",
  "asyncpg>=0.30.0",
  "bcrypt>=4.3.0",
  "celery>=5.4.0",
  "email-validator>=2.2.0",
  "fastapi>=0.115.12",
  "gunicorn>=23.0.0",
  "neo4j>=5.28.1",
  "PyJWT>=2.10.1",
  "prometheus-client>=0.21.1",
  "pydantic>=2.11.2",
  "pydantic-settings>=2.8.1",
  "psycopg[binary]>=3.2.6",
  "redis>=5.2.1",
  "sqlalchemy>=2.0.40",
  "structlog>=25.2.0",
  "uvicorn[standard]>=0.34.0",
]

[project.optional-dependencies]
dev = [
  "aiosqlite>=0.21.0",
  "httpx>=0.28.1",
  "pytest>=8.3.5",
  "pytest-asyncio>=0.26.0",
  "ruff>=0.11.2",
]

[tool.hatch.build.targets.wheel]
packages = ["src/csre"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N"]

[tool.ruff.lint.isort]
known-first-party = ["csre"]
